refactor: reconstruct loader components in accordance with atomic design
