+++ src/components/TextList.tsx
@@ -4,7 +4,7 @@ import TextComponent from './Text';
 import useNewTexts from '../hooks/useNewTexts';
 import useOldTexts from '../hooks/useOldTexts';
 import { maxPage } from '../lib/const';
-import LoadingIndicator from './LoadingIndicator';
+import LoadingIndicator from './molecules/LoadingIndicator';
 
 type Props = {
   userId?: string;
