+++ src/components/atoms/Spinner.tsx
@@ -1,0 +1,9 @@
+import { VscLoading } from 'react-icons/vsc';
+
+type Props = {
+  className: string;
+};
+
+export default function Spinner({ className }: Props) {
+  return <VscLoading className={`animate-spin ${className}`} />;
+}
